Moved shooting functionality out of Player class.
