+++ src/main/java/entity/Player.java
@@ -3,7 +3,7 @@ package entity;
 import display.DisplayEntity;
 import game.Audio;
 import game.Logger;
-import entity.builders.BulletBuilder;
+import entity.shooters.PlayerShooter;
 
 import java.util.List;
 import java.util.Random;
@@ -16,21 +16,15 @@ import java.util.Random;
 public class Player extends AbstractEntity {
 	private int lives;
 	private double rotation;
-	private long lastShot;
 	private long invincibleStart;
 	private int invincibleMS;
 	private long hyperspaceStart;
 	private boolean boost;
 	private boolean playerTwo;
-	private int maxBullets;
-	private double fireRate;
-	private int piercing;
 	private int shielding;
-	private boolean tripleShot;
-	private float bulletSize;
 	private int changeOfDying;
 	private String playerString;
-	private final BulletBuilder bBuilder;
+	private PlayerShooter shooter;
 
 	private static final int STARTING_LIVES = 3;
 	private static final float RADIUS = 5;
@@ -39,14 +33,9 @@ public class Player extends AbstractEntity {
 	private static final double ACCELERATION = .04;
 	private static final float DECELERATION = .01f;
 	private static final int HYPERSPACE_TIME = 1000;
-	private static final long FIRE_RATE = 200;
-	private static final float BULLET_SPEED = 4;
 	private static final float MAX_SPEED = 4;
-	private static final int MAX_BULLETS = 4;
 	private static final int CHANCE_OF_DYING = 25;
-	private static final float BULLET_SIZE = 2;
 	private static final float SPAWN_OFFSET = 40;
-    private static final double TRIPLE_SHOT_ANGLE = .1;
     
     /**
      * create an uninitialized player with only the default player.
@@ -59,17 +48,9 @@ public class Player extends AbstractEntity {
     	playerTwo = false;
     	playerString = "The player";
     	makeInvincible(INVINCIBILITY_START_TIME);
-		maxBullets = MAX_BULLETS;
-		fireRate = FIRE_RATE;
-		piercing = 1;
 		shielding = 0;
-		bulletSize = BULLET_SIZE;
-		tripleShot = false;
 		changeOfDying = CHANCE_OF_DYING;
-		// Initialize the Bullet Builder
-		bBuilder = new BulletBuilder();
-		bBuilder.setPierce(piercing);
-		bBuilder.setFriendly(true);
+		shooter = new PlayerShooter(this);
     }
 
 	/**
@@ -175,7 +156,7 @@ public class Player extends AbstractEntity {
 		}
 
 		if (input.contains("SPACE")) {
-			fire();
+			shooter.shoot();
 		}
 	}
 	
@@ -222,7 +203,7 @@ public class Player extends AbstractEntity {
 			}
 
 			if (input.contains("SPACE")) {
-				fire();
+				shooter.shoot();
 			}
 		}
 	}
@@ -252,7 +233,7 @@ public class Player extends AbstractEntity {
 		}
 
 		if (input.contains("ENTER")) {
-			fire();
+			shooter.shoot();
 		}
 	}
 
@@ -329,43 +310,6 @@ public class Player extends AbstractEntity {
 		hyperspaceStart = System.currentTimeMillis();
 		getThisGame().getAudio().play(Audio.TELEPORT);
 		}
-	}
-
-	/**
-	 * Method to handle firing bullets.
-	 */
-	private void fire() {
-		if (System.currentTimeMillis() - lastShot >	fireRate && getThisGame().bullets(this) < maxBullets) {
-			fireBullet(rotation);
-			if (tripleShot) {
-				fireBullet(rotation - TRIPLE_SHOT_ANGLE);
-				fireBullet(rotation + TRIPLE_SHOT_ANGLE);
-			}
-			lastShot = System.currentTimeMillis();
-			if (isPlayerTwo()) {
-				getThisGame().getAudio().playMultiple(Audio.SHOOTING2);
-			} else {
-				getThisGame().getAudio().playMultiple(Audio.SHOOTING);
-			}
-		}
-	}
-
-	/**
-	 * fire a bullet in a direction.
-	 * @param direction - the direction
-	 */
-	private void fireBullet(final double direction) {
-		bBuilder.setX(getX());
-		bBuilder.setY(getY());
-		bBuilder.setDX((float) (getDX() / 2 + Math.cos(direction) * BULLET_SPEED));
-		bBuilder.setDY((float) (getDY() / 2 - Math.sin(direction) * BULLET_SPEED));
-		bBuilder.setRadius(bulletSize);
-		bBuilder.setThisGame(getThisGame());
-		bBuilder.setShooter(this);
-		bBuilder.setPierce(piercing);
-		final Bullet b = (Bullet) bBuilder.getResult();
-		
-		getThisGame().create(b);
 	}
 
 	/**
@@ -511,94 +455,10 @@ public class Player extends AbstractEntity {
 	}
 
 	/**
-	 * @param bulletSize the bulletSize to set
-	 */
-	public final void setBulletSize(final float bulletSize) {
-		this.bulletSize = bulletSize;
-	}
-	
-	/**
-	 * @return bulletSize
-	 */
-	public final float getCurrentBulletSize() {
-		return bulletSize;
-	}
-	
-	/**
-	 * @param tripleShot the tripleShot to set
-	 */
-	public final void setTripleShot(final boolean tripleShot) {
-		this.tripleShot = tripleShot;
-	}
-
-	/**
-	 * @return the tripleShot
-	 */
-	public final boolean isTripleShot() {
-		return tripleShot;
-	}
-
-	/**
-	 * @param piercing the piercing to set
-	 */
-	public final void setPiercing(final int piercing) {
-		this.piercing = piercing;
-	}
-
-	/**
-	 * @return the piercing
-	 */
-	public final int getPiercing() {
-		return piercing;
-	}
-
-	/**
-	 * @param fireRate the fireRate to set
-	 */
-	public final void setFireRate(final double fireRate) {
-		this.fireRate = fireRate;
-	}
-	
-	/**
-	 * @return fireRate
-	 */
-	public final double getCurrentFireRate() {
-		return fireRate;
-	}
-
-	/**
-	 * @return the bulletSize
-	 */
-	public static float getBulletSize() {
-		return BULLET_SIZE;
-	}
-
-	/**
-	 * @return the fireRate
-	 */
-	public static long getFireRate() {
-		return FIRE_RATE;
-	}
-
-	/**
 	 * gain a shield level.
 	 */
 	public final void gainShield() {
 		shielding++;
-	}
-
-	/**
-	 * @param maxBullets the maxBullets to set
-	 */
-	public final void setMaxBullets(final int maxBullets) {
-		this.maxBullets = maxBullets;
-	}
-
-	/**
-	 * @return the maxBullets
-	 */
-	public static int getMaxBullets() {
-		return MAX_BULLETS;
 	}
 
 	/**
@@ -613,13 +473,6 @@ public class Player extends AbstractEntity {
 	 */
 	public final void setHyperspaceStart(final long hyperspaceStart) {
 		this.hyperspaceStart = hyperspaceStart;
-	}
-
-	/**
-	 * @return the lastShot
-	 */
-	public final long getLastShot() {
-		return lastShot;
 	}
 
 	/**
@@ -641,6 +494,13 @@ public class Player extends AbstractEntity {
 	 */
 	public final String getPlayerString() {
 		return playerString;
+	}
+	
+	/**
+	 * @return the Player Shooter
+	 */
+	public final PlayerShooter getShooter() {
+		return shooter;
 	}
 	
 	/**
