+++ src/main/java/entity/shooters/AbstractShooter.java
@@ -51,6 +51,13 @@ public abstract class AbstractShooter {
 	}
 	
 	/**
+	 * @param bBuilder the new bulletbuilder
+	 */
+	public final void setBBuilder(final BulletBuilder bBuilder) {
+		this.bBuilder = bBuilder;
+	}
+	
+	/**
 	 * @return the owner
 	 */
 	public final AbstractEntity getOwner() {
